+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### To Release(v0.3.0)
 
-* [vc] add ready and changes view block and reduce space between items
+* [vc] let changed files clickable, when clicked means ready to commit, else is not
+* [vc] remove ready/changes zone concept
 * [starter:fix] non-existed project would popup warning message and won't open IDE
 * [repl] switching expression from evaluated history
     * up select previous